pyproject.toml
[tool.poetry]
name = "service-area-api"
version = "0.1.0"
description = ""
authors = ["Artem <mitay93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
gunicorn = "^22.0.0"
djangorestframework-gis = "^1.0"
environs = "^11.0.0"
psycopg = "^3.1.19"
drf-spectacular = "^0.27.2"
drf-extensions = "^0.7.1"


[tool.poetry.group.test.dependencies]
faker = "^25.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
